* nsfmongo.c: Tcl-9 compat reform, adopt TCL_SIZE_T to avoid stack corruption in at least one function (NsfMongoCollectionQuery)

+++ library/mongodb/nsfmongo.c
@@ -1564,7 +1564,9 @@ NsfMongoGridFileCreate(Tcl_Interp *interp,
         iov.iov_len = (size_t)n;
         n = mongoc_gridfs_file_writev(gridFile, &iov, 1, 0);
         if ((size_t)n != iov.iov_len) {
-          NsfLog(interp, NSF_LOG_WARN, "mongodb: write of %zu bytes returned %zu", iov.iov_len, n);
+          NsfLog(interp, NSF_LOG_WARN,
+                 "mongodb: write of %zu bytes returned %lld",
+                 iov.iov_len, (long long)n);
         }
       } else if (n == 0) {
         break;
